M2kHardwareTrigger: Move the M2kHardwareTrigger from namespace "libm2k.analog" to namespace "libm2k".
The M2kHardwareTrigger class handles both analog and digital trigger,
so keeping it in the "analog" namespace is confusing.
