+++ tools/m2kcli/commands/analog/analog_in.cpp
@@ -273,7 +273,7 @@ void AnalogIn::handleSet()
 				throw std::runtime_error("Invalid trigger source: '" + std::string(value) + "'.\n");
 			}
 			analogIn->getTrigger()->setAnalogSource(
-				static_cast<libm2k::analog::M2K_TRIGGER_SOURCE_ANALOG>(enumIndex));
+				static_cast<libm2k::M2K_TRIGGER_SOURCE_ANALOG>(enumIndex));
 		} else if (argument == "trigger_delay") {
 			analogIn->getTrigger()->setAnalogDelay(std::stoi(value));
 		} else if (argument == "kernel_buffers") {
@@ -320,7 +320,7 @@ void AnalogIn::handleSetChannel()
 			}
 			for (int &channel : channels) {
 				analogIn->getTrigger()->setAnalogCondition(channel,
-									   static_cast<libm2k::analog::M2K_TRIGGER_CONDITION_ANALOG>(enumIndex));
+									   static_cast<libm2k::M2K_TRIGGER_CONDITION_ANALOG>(enumIndex));
 			}
 		} else if (argument == "trigger_mode") {
 			int enumIndex = getIndexOfElement(value, triggerMode);
@@ -329,7 +329,7 @@ void AnalogIn::handleSetChannel()
 			}
 			for (int &channel : channels) {
 				analogIn->getTrigger()->setAnalogMode(channel,
-								      static_cast<libm2k::analog::M2K_TRIGGER_MODE>(enumIndex));
+								      static_cast<libm2k::M2K_TRIGGER_MODE>(enumIndex));
 			}
 		} else if (argument == "trigger_hysteresis") {
 			for (int &channel : channels) {
